refactor: 💡 remove old models
