+++ jest.config.js
@@ -6,7 +6,13 @@ module.exports = {
     '.+\\.(css|styl|less|sass|scss)$': `identity-obj-proxy`,
     // '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': `<rootDir>/__mocks__/file-mock.js`,
   },
-  testPathIgnorePatterns: [`node_modules`, `.cache`, 'build', 'assets'],
+  testPathIgnorePatterns: [
+    `node_modules`,
+    `.cache`,
+    'build',
+    'assets',
+    'starters',
+  ],
   transformIgnorePatterns: [`node_modules/(?!(gatsby)/)`],
   modulePaths: ['<rootDir>/'],
   globals: {
