+++ packages/mui-layout/src/models/edgeSidebar.js
@@ -1,15 +1,6 @@
 export default (ctx = {}) => {
   const { sidebar = {}, collapsed } = ctx;
 
-  if (
-    Object.keys(sidebar).length === 0 &&
-    process.env.NODE_ENV !== 'production'
-  ) {
-    console.warn(
-      'Sidebar is empty! please make sure that you pass { sidebar: {...} } to config'
-    );
-  }
-
   const getSidebarWidth = () => {
     if (sidebar.hidden) return 0;
     return sidebar.collapsible && collapsed
