+++ packages/mui-layout/src/models/footer.js
@@ -1,34 +1,40 @@
-import createSidebarEffect from './sidebarEffect';
+import createAllSidebars from './allSidebars';
+import { reduceWidths } from './width';
 
 export default (ctx = {}) => {
-  const { sidebar = {}, secondarySidebar = {}, footer = {} } = ctx;
-  const { getStyle } = createSidebarEffect(ctx, footer);
+  const { footer = {} } = ctx;
+  const {
+    mainEffect,
+    subEffect,
+    mapSecondaryConfig,
+    isEdgeAndInset,
+    isPrimaryInset,
+    isPrimaryEdge,
+    isSecondaryEdge,
+    isSecondaryInset,
+  } = createAllSidebars(ctx);
+  const subFooter = mapSecondaryConfig(footer);
+  const mainAttached = !isEdgeAndInset || (isPrimaryInset && isSecondaryEdge);
+  const subAttached = !isEdgeAndInset || (isPrimaryEdge && isSecondaryInset);
   return {
-    getStyle: () => {
-      if (sidebar.inset && secondarySidebar.inset) {
-        return {
-          ...getStyle({
-            primaryDisabled: footer.insetBehavior !== 'fit',
-            secondaryDisabled: footer.secondaryInsetBehavior !== 'fit',
-          }),
-          width: 'auto',
-        };
-      }
-      if (sidebar.inset && !secondarySidebar.inset) {
-        return {
-          ...getStyle({
-            primaryDisabled: footer.insetBehavior !== 'fit',
-          }),
-        };
-      }
-      if (!sidebar.inset && secondarySidebar.inset) {
-        return {
-          ...getStyle({
-            secondaryDisabled: footer.secondaryInsetBehavior !== 'fit',
-          }),
-        };
-      }
-      return getStyle();
+    getMarginStyle() {
+      return {
+        ...(mainAttached && mainEffect.getMarginStyle(footer)),
+        ...(subAttached && subEffect.getMarginStyle(subFooter)),
+      };
+    },
+    getWidthStyle() {
+      const finalWidthObj = reduceWidths([
+        ...(mainAttached ? [mainEffect.getWidthObj(footer)] : []),
+        ...(subAttached ? [subEffect.getWidthObj(subFooter)] : []),
+      ]);
+      return finalWidthObj.getStyle();
+    },
+    getStyle() {
+      return {
+        ...this.getMarginStyle(),
+        ...this.getWidthStyle(),
+      };
     },
   };
 };
