+++ packages/mui-layout/src/utils/index.js
@@ -1,4 +1,3 @@
-export { default as createLayoutUtils } from './LayoutUtils';
 export { default as isSomeExisted } from './isSomeExisted';
 export { default as selectConfigByScreen } from './selectConfigByScreen';
 export { default as getInitialScreen } from './getInitialScreen';
